# [et] Make native-unit-tests command faster and fix reliability